FuturePath AI/requirements.txt
openai>=1.1.0
python-dotenv
streamlit
fpdf
